Rename GlobalUsings file to Usings v2 and some tidy up
