+++ tests/DotNet.Sdk.Extensions.Tests/Options/AddOptionsValue/Auxiliary/MyOptions.cs
@@ -1,0 +1,6 @@
+namespace DotNet.Sdk.Extensions.Tests.Options.AddOptionsValue.Auxiliary;
+
+internal sealed class MyOptions
+{
+    public string? SomeOption { get; set; }
+}
